Update codebase to SDL3, use float-based rendering
Migrated all code from SDL2/SDL2_image to SDL3/SDL3_image, updating includes, function calls, and event constants to their SDL3 equivalents. Refactored all physics, rendering, and coordinate code to use float types and SDL_FRect for consistency with SDL3's float-based API. Updated project files to use new include/library paths and SDL3 libraries. Adjusted window/renderer creation and event handling for SDL3. Cleaned up redundant code and updated solution/project structure. Added new Visual Studio database/indexing files.

+++ 3_PhaseTransition/GameByFunctions/GameFuncStage1.cpp
@@ -7,16 +7,16 @@ static SDL_FRect g_destination_rectangle_stage1; // for destination
 
 static bool g_character_go_left;
 static bool g_character_go_right;
-static double g_character_x;
-static double g_character_y;
+static float g_character_x;
+static float g_character_y;
 
 void Init_Stage1()
 {
 	// For Character
 	g_character_go_left = false;
 	g_character_go_right = false;
-	g_character_x = 300;
-	g_character_y = 300;
+	g_character_x = 300.0f;
+	g_character_y = 300.0f;
 
 
 	// For Texture
@@ -28,8 +28,8 @@ void Init_Stage1()
 	SDL_GetTextureSize(g_texture_stage1, &g_source_rectangle_stage1.w, &g_source_rectangle_stage1.h);
 	g_source_rectangle_stage1.x = 0;
 	g_source_rectangle_stage1.y = 0;
-	g_destination_rectangle_stage1.x = (int)g_character_x; 
-	g_destination_rectangle_stage1.y = (int)g_character_y; 
+	g_destination_rectangle_stage1.x = g_character_x; 
+	g_destination_rectangle_stage1.y = g_character_y; 
 	g_destination_rectangle_stage1.w = g_source_rectangle_stage1.w;
 	g_destination_rectangle_stage1.h = g_source_rectangle_stage1.h;
 
@@ -41,13 +41,13 @@ void Update_Stage1()
 	if ( g_character_go_left )
 	{
 		g_character_x -= 10.;
-		g_destination_rectangle_stage1.x = (int)g_character_x;
+		g_destination_rectangle_stage1.x = g_character_x;
 	}
 
 	if ( g_character_go_right )
 	{
 		g_character_x += 10.;
-		g_destination_rectangle_stage1.x = (int)g_character_x;
+		g_destination_rectangle_stage1.x = g_character_x;
 	}
 }
